actions: Update dotnet.yml
Add support for C/C++ in the pipeline and upload the newer igcs connector.

It was necessary to create two different build scripts because for some reason
when you install the msvc compiler in a github action everything Just Breaks.

Then we create two different jobs to make sure they're independent instances of
containers that don't mess with other's environments.